+++ public/index.html
@@ -10,10 +10,6 @@
             content="Web site created using create-react-app"
         />
         <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-        <!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
         <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
         <style>
             * {
